refactor get included projects to new class
