+++ Typewriter.CLI/Typewriter.cs
@@ -39,49 +39,25 @@ namespace Typewriter.CLI
                 };
                 var template = new Template(templateInfo);
                 var settings = (template.Settings as SettingsImpl);
-                var includedProjects = settings.IncludedProjects;
+
+                ICollection<string> includedProjects;
+                try
+                {
+                    var provider = new IncludedProjectsProvider(logger);
+                    includedProjects = provider.GetIncludedProjects(settings, projectPath, projects);
+                }
+                catch (ArgumentException e)
+                {
+                    logger.LogError(e.Message);
+                    continue;
+                }
+                
 
                 if (projects == null)
                 {
-                    if (settings.ShouldIncludeAllProjects)
-                    {
-                        includedProjects = null; // include all projects
-                    }
-                    if (settings.ShouldIncludeCurrentProject && !settings.ShouldIncludeAllProjects)
-                    {
-                        logger.LogError("IncludeCurrentProject is not supported. Please use IncludeProject");
-                        return;
-                    }
-                    if (settings.ShouldIncludeReferencedProjects)
-                    {
-                        logger.LogError("IncludeReferencedProjects is not supported. Please use IncludeProject");
-                        return;
-                    }
                     projects = solution.GetProjects(includedProjects);
                 }
-                else
-                {
-                    if (settings.ShouldIncludeCurrentProject)
-                    {
-                        logger.LogWarning("IncludeCurrentProject is not supported. Project provided in projectPath parameter will be used");
-                        var currentProject = GetCurrentProject(projectPath, projects);
-                        includedProjects.Add(currentProject.Name);
-                    }
-                    if (settings.ShouldIncludeReferencedProjects)
-                    {
-                        logger.LogWarning("IncludeReferencedProjects is not supported. Referenced Projects for Project provided in projectPath parameter will be used");
-                        var currentProject = GetCurrentProject(projectPath, projects);
-                        foreach (var referencedProject in GetReferencedProjects(projects, currentProject))
-                        {
-                            includedProjects.Add(referencedProject.Name);
-                        }
-                    }
-                    if (settings.ShouldIncludeAllProjects)
-                    {
-                        logger.LogError("IncludeAllProjects is not supported when project parameter is provided");
-                        return;
-                    }
-                }
+
                 RenderTemplateForAllProjects(template, projects.Where(p => includedProjects == null || includedProjects.Contains(p.Name)));
             }
 
@@ -89,31 +65,22 @@ namespace Typewriter.CLI
             logger.LogInformation($"Done in {TimeSpan.FromMilliseconds(globalStopWatch.ElapsedMilliseconds).TotalSeconds} seconds");
         }
 
-        private static IEnumerable<Project> GetReferencedProjects(IEnumerable<Project> projects, Project currentProject)
-        {
-            var referencesId = currentProject.ProjectReferences.Select(r => r.ProjectId);
-            var referencedProjects = projects.Where(p => referencesId.Contains(p.Id));
-            return referencedProjects;
-        }
-
-        private static Project GetCurrentProject(string projectPath, IEnumerable<Project> projects)
-        {
-            return projects.Where(p => p.FilePath.Equals(projectPath, StringComparison.InvariantCultureIgnoreCase)).Single();
-        }
 
         private void RenderTemplateForAllProjects(Template template, IEnumerable<Project> projects)
         {
             foreach (var project in projects)
             {
                 Stopwatch stopwatch = Stopwatch.StartNew();
-                logger.LogInformation($"Rendering template {template.TemplatePath}");
+                logger.LogInformation($"Rendering template {template.TemplatePath} for project {project.Name}");
+
                 RenderProject(template, project);
+
                 stopwatch.Stop();
                 logger.LogDebug($"Rendered in {TimeSpan.FromMilliseconds(stopwatch.ElapsedMilliseconds).TotalSeconds} seconds");
             }
         }
 
-        private static void RenderProject(Template template, Microsoft.CodeAnalysis.Project project)
+        private void RenderProject(Template template, Project project)
         {
             foreach (var projectFile in project.Documents)
             {
